refactor(launch/manager): split input.rs into stage-focused submodules
The single-file `input.rs` grew to ~4750 lines and combined top-level
dispatch, list/editor/prelude/confirm-delete stage handlers, save-flow
plumbing, mouse/geometry helpers, and a large test suite. Split it
into `input/` as a directory module:

- `input/mod.rs` — thin top-level dispatcher (`handle_key`, `handle_mouse`),
  `InputOutcome`, re-exports.
- `input/list.rs` — list-stage key/modal handling.
- `input/editor.rs` — editor-stage key/modal handling, tab dispatch,
  agent-allow toggles, agent-default selection.
- `input/save.rs` — `begin_editor_save`, `commit_editor_save`, save-flow
  helpers (`build_workspace_edit`, confirm-save line builders, etc.).
- `input/prelude.rs` — create-workspace wizard.
- `input/mouse.rs` — mouse events, seam drag, list hit-test, file-browser
  URL click.

Tests travel with the behavior they exercise. No runtime behavior
change — key bindings, mouse behavior, save semantics, and dispatch
order are all byte-identical.

Addresses finding #6 of the PR #166 current-branch review.
